Moved page to manage user images to a new route called `/manage` and made `User Profile Page` look more similar to our `Explore Page`
