+++ app/pages/ManageImagesPage/components/DeleteImageButton/DeleteImageButton.tsx
