+++ app/pages/ManageImagesPage/components/DeleteImageButton/index.ts
