+++ app/pages/ManageImagesPage/components/EditImageButton/EditImageButton.tsx
