+++ app/pages/ManageImagesPage/components/EditImageButton/index.ts
