+++ app/pages/ManageImagesPage/components/ToggleIsImagePrivateButton/ToggleIsImagePrivateButton.tsx
