+++ app/pages/ManageImagesPage/components/ToggleIsImagePrivateButton/index.ts
