+++ app/pages/ManageImagesPage/index.ts
@@ -1,0 +1,1 @@
+export { default as ManageImagesPage } from "./ManageImagesPage";
