+++ app/pages/index.ts
@@ -3,5 +3,6 @@ export * from "./LandingPage";
 export * from "./ExplorePage";
 export * from "./CollectionsPage";
 export * from "./UserProfilePage";
+export * from "./ManageImagesPage";
 export * from "./CollectionDetailsPage";
 export * from "./SettingsPage";
