+++ app/routes/manage.tsx
@@ -7,7 +7,7 @@ import {
 } from "@remix-run/node";
 import { isRouteErrorResponse, useRouteError } from "@remix-run/react";
 import { Alert } from "antd";
-import { UserProfilePage } from "~/pages";
+import { ManageImagesPage } from "~/pages";
 import { getUserData } from "~/server";
 import { getSession } from "~/services";
 import { authenticator } from "~/services/auth.server";
@@ -27,22 +27,21 @@ export const meta: MetaFunction<
     userMatch?.data.data?.username || userMatch?.data.data?.name || userId;
 
   return [
-    { title: `${username} | Profile` },
+    { title: `${username} | Manage Images` },
     {
-      name: "description",
-      content: `Checkout ${username}'s AI generated images`,
+      name: "Manage AI generated images",
+      content: "Manage AI generated images",
     },
   ];
 };
 
 export const loader = async ({ request, params }: LoaderFunctionArgs) => {
-  const userId = params.userId || "";
-  invariantResponse(userId, "User does not exist");
-
   const session = await getSession(request.headers.get("Cookie"));
   const googleSessionData = (await session.get("_session")) || undefined;
   const currentLoggedInUserData = googleSessionData;
   const currentLoggedInUserID = currentLoggedInUserData.id || "";
+
+  invariantResponse(currentLoggedInUserID, "User does not exist");
 
   await authenticator.isAuthenticated(request, {
     failureRedirect: "/",
@@ -50,12 +49,12 @@ export const loader = async ({ request, params }: LoaderFunctionArgs) => {
 
   const searchParams = new URL(request.url).searchParams;
   const currentPage = Math.max(Number(searchParams.get("page") || 1), 1);
-  const pageSize = Number(searchParams.get("page_size")) || 250;
+  const pageSize = Number(searchParams.get("page_size")) || 50;
 
   // If UserA is visiting UserB's profile, we do not want to show UserB's Private images to UserA
-  const shouldGetUsersPrivateImages = userId === currentLoggedInUserID;
+  const shouldGetUsersPrivateImages = true;
   const data = await getUserData(
-    userId,
+    currentLoggedInUserID,
     currentPage,
     pageSize,
     shouldGetUsersPrivateImages,
@@ -66,7 +65,7 @@ export const loader = async ({ request, params }: LoaderFunctionArgs) => {
   return json(data);
 };
 
-export type UserProfilePageLoader = SerializeFrom<typeof loader>;
+export type ManageImagesPageLoader = SerializeFrom<typeof loader>;
 
 // export async function action({ request }: ActionFunctionArgs) {
 //   const formData = await request.formData();
@@ -90,7 +89,7 @@ export type UserProfilePageLoader = SerializeFrom<typeof loader>;
 // }
 
 export default function Index() {
-  return <UserProfilePage />;
+  return <ManageImagesPage />;
 }
 
 export function ErrorBoundary() {
